% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{paperplanes}
\alias{paperplanes}
\title{Paper Planes.}
\format{A data frame with eight variables:
\code{id},
\code{hour} (the hour of the day),
\code{person},
\code{gender},
\code{age},
\code{plane},
\code{paper} (two types: 80g and 120g),
\code{distance} (recorded in meter).}
\usage{
paperplanes
}
\description{
Distance Recordings from a Paper Plane Folding/Flying Experiment
}
\details{
Visitors of a science fair (the 3. 'Nacht des Wissens' in
Goettingen/Germany \url{http://www.ndw.uni-goettingen.de/}) had the
opportunity to throw a paper plane -- either folding one
themselves or re-using an existing one.
}
\keyword{datasets}

